test for new w/.png + db.sqlite3 added + entradas a la db. Project readygit add .
