+++ .gitignore
@@ -5,7 +5,7 @@ venv/*
 *.pyc
 */__pycache__/
 x_*
-db.sqlite3
+
 ˝
 static/*
 staticfiles/*
