+++ README.md
@@ -3,10 +3,10 @@
 ## Participantes: Agustin D. Gibertoni + Fabricio Jerez
 
 
-- Las apps vinculadas al home, picture (galería) y los módulos de login/signin/logout y manejo de perfiles fueron desarrolladas por Agustín.
+- Las apps vinculadas al home, picture (galería), about, y módulos de login/signin/logout y manejo de perfiles fueron desarrolladas por Agustín.
 
 - Las apps vinculadas a top10 y news (novedades) fueron desarroladas por Fabricio.
 
-- En conjunto hicimos la parte del About. `Todo el código fue chequeado por ambos integrantes y testeado su correcto funcionamiento.`
+- `Todo el código fue chequeado por ambos integrantes y testeado su correcto funcionamiento.`
 
 - La creación de contenido (entradas / sql db) también fue hecha en conjunto a través de nuestros superusers u otros usuarios q creamos extra (o desde el admin), utilizando información recolectada de diferentes sitios web.
